latest version of graphics created (still bugs)
This graphics version is the latest but has a lot of bugs and things that need to be dealt with. First, while loop in virtualPong (might need to be removed), time.sleep and pop in drawCup (to test if cups pop), and most of cupHit collision code currently commented out because it needs to run when the data is received from opencv. And in drawBall it breaks the loop when yPos is below 0.